[REFACTORING] Organized Project Repo
