+++ extension/packagemanagerextension/serverextension/serverextension.py
@@ -7,8 +7,7 @@ from notebook.utils import url_path_join as ujoin
 
 from .envmanager import EnvManager
 
-from .projectmanagement import *
-from .packagemanagement import *
+from .apihandlers import *
 
 NS = r'api/packagemanager'
 
@@ -29,7 +28,7 @@ default_handlers = [
 
 
 def PackageManagerHandler(nbapp):
-    """Load the nbserver extension"""
+    '''Load the nbserver extension'''
     webapp = nbapp.web_app
     webapp.settings['env_manager'] = EnvManager(parent=nbapp)
